+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Time signature in tab had staff's color.
 - Typo MusicErrorType.Timesignature => MusicErrorType.TimeSignature.
 - Set play state to Stopped after last note stopped playing (was when last note started playing).
+- Improved way to compute and spread rhythm columns in measures. Fixes unevenly spaced columns.
 
 ## [6.3.7] - 2026-02-05
 ### Fixed